COMMERCE-6745 Split portlet and PortletSharedSearchContributor to avoid circular dependencies.
