+++ modules/apps/commerce/commerce-product-content-search-web/src/main/java/com/liferay/commerce/product/content/search/web/internal/portlet/CPOptionFacetsPortlet.java
@@ -14,47 +14,29 @@
 
 package com.liferay.commerce.product.content.search.web.internal.portlet;
 
-import com.liferay.asset.kernel.model.AssetCategory;
-import com.liferay.commerce.account.model.CommerceAccount;
-import com.liferay.commerce.account.util.CommerceAccountHelper;
 import com.liferay.commerce.product.constants.CPField;
 import com.liferay.commerce.product.constants.CPPortletKeys;
 import com.liferay.commerce.product.content.search.web.internal.display.context.CPOptionFacetsDisplayContext;
 import com.liferay.commerce.product.content.search.web.internal.util.CPOptionFacetsUtil;
-import com.liferay.commerce.product.model.CPDefinition;
 import com.liferay.commerce.product.model.CPOption;
-import com.liferay.commerce.product.model.CommerceChannel;
 import com.liferay.commerce.product.service.CPOptionLocalService;
-import com.liferay.commerce.product.service.CommerceChannelLocalService;
-import com.liferay.commerce.search.facet.SerializableFacet;
-import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
-import com.liferay.portal.kernel.search.Indexer;
-import com.liferay.portal.kernel.search.IndexerRegistryUtil;
-import com.liferay.portal.kernel.search.QueryConfig;
-import com.liferay.portal.kernel.search.SearchContext;
 import com.liferay.portal.kernel.search.facet.Facet;
-import com.liferay.portal.kernel.search.facet.MultiValueFacet;
-import com.liferay.portal.kernel.search.facet.SimpleFacet;
 import com.liferay.portal.kernel.search.facet.collector.FacetCollector;
 import com.liferay.portal.kernel.search.facet.collector.TermCollector;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.search.searcher.SearchRequest;
 import com.liferay.portal.search.searcher.SearchResponse;
-import com.liferay.portal.search.web.portlet.shared.search.PortletSharedSearchContributor;
 import com.liferay.portal.search.web.portlet.shared.search.PortletSharedSearchRequest;
 import com.liferay.portal.search.web.portlet.shared.search.PortletSharedSearchResponse;
-import com.liferay.portal.search.web.portlet.shared.search.PortletSharedSearchSettings;
 
 import java.io.IOException;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 import javax.portlet.Portlet;
 import javax.portlet.PortletException;
@@ -88,100 +70,9 @@ import org.osgi.service.component.annotations.Reference;
 		"javax.portlet.resource-bundle=content.Language",
 		"javax.portlet.security-role-ref=guest,power-user,user"
 	},
-	service = {Portlet.class, PortletSharedSearchContributor.class}
+	service = Portlet.class
 )
-public class CPOptionFacetsPortlet
-	extends MVCPortlet implements PortletSharedSearchContributor {
-
-	@Override
-	public void contribute(
-		PortletSharedSearchSettings portletSharedSearchSettings) {
-
-		RenderRequest renderRequest =
-			portletSharedSearchSettings.getRenderRequest();
-
-		try {
-			SearchContext searchContext =
-				portletSharedSearchSettings.getSearchContext();
-
-			SerializableFacet serializableFacet = new SerializableFacet(
-				CPField.OPTION_NAMES, searchContext);
-
-			Optional<String[]> parameterValuesOptional =
-				portletSharedSearchSettings.getParameterValues71(
-					CPField.OPTION_NAMES);
-
-			if (parameterValuesOptional.isPresent()) {
-				serializableFacet.select(parameterValuesOptional.get());
-
-				searchContext.setAttribute(
-					CPField.OPTION_NAMES, parameterValuesOptional.get());
-			}
-
-			portletSharedSearchSettings.addFacet(serializableFacet);
-
-			List<Facet> facets = getFacets(renderRequest);
-
-			for (Facet facet : facets) {
-				String cpOptionKey =
-					CPOptionFacetsUtil.getCPOptionKeyFromIndexFieldName(
-						facet.getFieldName());
-
-				parameterValuesOptional =
-					portletSharedSearchSettings.getParameterValues71(
-						cpOptionKey);
-
-				serializableFacet = new SerializableFacet(
-					facet.getFieldName(), searchContext);
-
-				if (parameterValuesOptional.isPresent()) {
-					serializableFacet.select(parameterValuesOptional.get());
-
-					searchContext.setAttribute(
-						facet.getFieldName(), parameterValuesOptional.get());
-				}
-
-				portletSharedSearchSettings.addFacet(serializableFacet);
-			}
-
-			ThemeDisplay themeDisplay =
-				(ThemeDisplay)renderRequest.getAttribute(WebKeys.THEME_DISPLAY);
-
-			long commerceChannelGroupId = 0;
-
-			CommerceChannel commerceChannel =
-				_commerceChannelLocalService.fetchCommerceChannelBySiteGroupId(
-					themeDisplay.getScopeGroupId());
-
-			if (commerceChannel != null) {
-				commerceChannelGroupId = commerceChannel.getGroupId();
-			}
-
-			if (commerceChannelGroupId > 0) {
-				searchContext.setAttribute(
-					CPField.CHANNEL_GROUP_ID, commerceChannelGroupId);
-				searchContext.setAttribute("secure", Boolean.TRUE);
-
-				CommerceAccount commerceAccount =
-					_commerceAccountHelper.getCurrentCommerceAccount(
-						commerceChannelGroupId, themeDisplay.getRequest());
-
-				long[] commerceAccountGroupIds = null;
-
-				if (commerceAccount != null) {
-					commerceAccountGroupIds =
-						_commerceAccountHelper.getCommerceAccountGroupIds(
-							commerceAccount.getCommerceAccountId());
-				}
-
-				searchContext.setAttribute(
-					"commerceAccountGroupIds", commerceAccountGroupIds);
-			}
-		}
-		catch (Exception exception) {
-			_log.error(exception, exception);
-		}
-	}
+public class CPOptionFacetsPortlet extends MVCPortlet {
 
 	@Override
 	public void render(
@@ -234,111 +125,6 @@ public class CPOptionFacetsPortlet
 		super.render(renderRequest, renderResponse);
 	}
 
-	protected SearchContext buildSearchContext(RenderRequest renderRequest)
-		throws PortalException {
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)renderRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		SearchContext searchContext = new SearchContext();
-
-		searchContext.setCompanyId(themeDisplay.getCompanyId());
-		searchContext.setLayout(themeDisplay.getLayout());
-		searchContext.setLocale(themeDisplay.getLocale());
-		searchContext.setTimeZone(themeDisplay.getTimeZone());
-		searchContext.setUserId(themeDisplay.getUserId());
-
-		QueryConfig queryConfig = searchContext.getQueryConfig();
-
-		queryConfig.setLocale(themeDisplay.getLocale());
-
-		searchContext.setAttribute(CPField.PUBLISHED, Boolean.TRUE);
-
-		CommerceChannel commerceChannel =
-			_commerceChannelLocalService.fetchCommerceChannelBySiteGroupId(
-				themeDisplay.getScopeGroupId());
-
-		if (commerceChannel != null) {
-			searchContext.setAttribute(
-				"commerceChannelGroupId", commerceChannel.getGroupId());
-
-			CommerceAccount commerceAccount =
-				_commerceAccountHelper.getCurrentCommerceAccount(
-					commerceChannel.getGroupId(),
-					_portal.getHttpServletRequest(renderRequest));
-
-			if (commerceAccount != null) {
-				long[] commerceAccountGroupIds =
-					_commerceAccountHelper.getCommerceAccountGroupIds(
-						commerceAccount.getCommerceAccountId());
-
-				searchContext.setAttribute(
-					"commerceAccountGroupIds", commerceAccountGroupIds);
-			}
-		}
-
-		searchContext.setAttribute("secure", Boolean.TRUE);
-
-		return searchContext;
-	}
-
-	protected List<Facet> getFacets(RenderRequest renderRequest)
-		throws PortalException {
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)renderRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		List<Facet> facets = new ArrayList<>();
-
-		AssetCategory assetCategory = (AssetCategory)renderRequest.getAttribute(
-			WebKeys.ASSET_CATEGORY);
-
-		SearchContext searchContext = buildSearchContext(renderRequest);
-
-		if (assetCategory != null) {
-			searchContext.setAssetCategoryIds(
-				new long[] {assetCategory.getCategoryId()});
-		}
-
-		Facet facet = new SimpleFacet(searchContext);
-
-		facet.setFieldName(CPField.OPTION_NAMES);
-
-		searchContext.addFacet(facet);
-
-		QueryConfig queryConfig = searchContext.getQueryConfig();
-
-		queryConfig.addSelectedFieldNames(CPField.OPTION_NAMES);
-
-		queryConfig.setHighlightEnabled(false);
-		queryConfig.setScoreEnabled(false);
-
-		Indexer<CPDefinition> indexer = IndexerRegistryUtil.nullSafeGetIndexer(
-			CPDefinition.class);
-
-		indexer.search(searchContext);
-
-		FacetCollector facetCollector = facet.getFacetCollector();
-
-		for (TermCollector termCollector : facetCollector.getTermCollectors()) {
-			CPOption cpOption = _cpOptionLocalService.getCPOption(
-				searchContext.getCompanyId(), termCollector.getTerm());
-
-			if (cpOption.isFacetable()) {
-				MultiValueFacet multiValueFacet = new MultiValueFacet(
-					searchContext);
-
-				multiValueFacet.setFieldName(
-					CPOptionFacetsUtil.getIndexFieldName(
-						termCollector.getTerm(), themeDisplay.getLanguageId()));
-
-				facets.add(multiValueFacet);
-			}
-		}
-
-		return facets;
-	}
-
 	protected String getPaginationStartParameterName(
 		PortletSharedSearchResponse portletSharedSearchResponse) {
 
@@ -357,15 +143,6 @@ public class CPOptionFacetsPortlet
 		CPOptionFacetsPortlet.class);
 
 	@Reference
-	private CommerceAccountHelper _commerceAccountHelper;
-
-	@Reference
-	private CommerceChannelLocalService _commerceChannelLocalService;
-
-	@Reference
 	private CPOptionLocalService _cpOptionLocalService;
-
-	@Reference
-	private Portal _portal;
 
 }
