+++ modules/apps/commerce/commerce-product-content-search-web/src/main/java/com/liferay/commerce/product/content/search/web/internal/portlet/CPPriceRangeFacetsPortlet.java
@@ -17,35 +17,18 @@ package com.liferay.commerce.product.content.search.web.internal.portlet;
 import com.liferay.commerce.currency.util.CommercePriceFormatter;
 import com.liferay.commerce.product.constants.CPField;
 import com.liferay.commerce.product.constants.CPPortletKeys;
-import com.liferay.commerce.product.content.search.web.internal.configuration.CPPriceRangeFacetsPortletInstanceConfiguration;
 import com.liferay.commerce.product.content.search.web.internal.display.context.CPPriceRangeFacetsDisplayContext;
-import com.liferay.commerce.search.facet.SerializableFacet;
-import com.liferay.petra.string.StringPool;
-import com.liferay.portal.json.JSONObjectImpl;
 import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.json.JSONFactoryUtil;
-import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
-import com.liferay.portal.kernel.search.SearchContext;
-import com.liferay.portal.kernel.search.facet.Facet;
-import com.liferay.portal.kernel.search.facet.RangeFacet;
-import com.liferay.portal.kernel.search.facet.config.FacetConfiguration;
-import com.liferay.portal.kernel.theme.PortletDisplay;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.search.searcher.SearchRequest;
 import com.liferay.portal.search.searcher.SearchResponse;
-import com.liferay.portal.search.web.portlet.shared.search.PortletSharedSearchContributor;
 import com.liferay.portal.search.web.portlet.shared.search.PortletSharedSearchRequest;
 import com.liferay.portal.search.web.portlet.shared.search.PortletSharedSearchResponse;
-import com.liferay.portal.search.web.portlet.shared.search.PortletSharedSearchSettings;
 
 import java.io.IOException;
-
-import java.util.Optional;
 
 import javax.portlet.Portlet;
 import javax.portlet.PortletException;
@@ -80,44 +63,9 @@ import org.osgi.service.component.annotations.Reference;
 		"javax.portlet.resource-bundle=content.Language",
 		"javax.portlet.security-role-ref=guest,power-user,user"
 	},
-	service = {Portlet.class, PortletSharedSearchContributor.class}
+	service = Portlet.class
 )
-public class CPPriceRangeFacetsPortlet
-	extends MVCPortlet implements PortletSharedSearchContributor {
-
-	@Override
-	public void contribute(
-		PortletSharedSearchSettings portletSharedSearchSettings) {
-
-		RenderRequest renderRequest =
-			portletSharedSearchSettings.getRenderRequest();
-
-		try {
-			SearchContext searchContext =
-				portletSharedSearchSettings.getSearchContext();
-
-			Facet facet = getFacet(renderRequest, searchContext);
-
-			Optional<String[]> parameterValuesOptional =
-				portletSharedSearchSettings.getParameterValues71(
-					facet.getFieldName());
-
-			SerializableFacet serializableFacet = new SerializableFacet(
-				facet.getFieldName(), searchContext);
-
-			if (parameterValuesOptional.isPresent()) {
-				serializableFacet.select(parameterValuesOptional.get());
-
-				searchContext.setAttribute(
-					facet.getFieldName(), parameterValuesOptional.get());
-			}
-
-			portletSharedSearchSettings.addFacet(facet);
-		}
-		catch (Exception exception) {
-			_log.error(exception, exception);
-		}
-	}
+public class CPPriceRangeFacetsPortlet extends MVCPortlet {
 
 	@Override
 	public void render(
@@ -145,46 +93,6 @@ public class CPPriceRangeFacetsPortlet
 		}
 
 		super.render(renderRequest, renderResponse);
-	}
-
-	protected Facet getFacet(
-			RenderRequest renderRequest, SearchContext searchContext)
-		throws PortalException {
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)renderRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		PortletDisplay portletDisplay = themeDisplay.getPortletDisplay();
-
-		CPPriceRangeFacetsPortletInstanceConfiguration
-			cpPriceRangeFacetsPortletInstanceConfiguration =
-				portletDisplay.getPortletInstanceConfiguration(
-					CPPriceRangeFacetsPortletInstanceConfiguration.class);
-
-		Facet facet = new RangeFacet(searchContext);
-
-		FacetConfiguration facetConfiguration = new FacetConfiguration();
-
-		JSONObject jsonObject = new JSONObjectImpl();
-
-		String rangesJSONArrayString =
-			cpPriceRangeFacetsPortletInstanceConfiguration.
-				rangesJSONArrayString();
-
-		rangesJSONArrayString = StringUtil.replace(
-			rangesJSONArrayString, new String[] {"\\,", StringPool.STAR},
-			new String[] {StringPool.COMMA, String.valueOf(Double.MAX_VALUE)});
-
-		jsonObject.put(
-			"ranges", JSONFactoryUtil.createJSONArray(rangesJSONArrayString));
-
-		facetConfiguration.setDataJSONObject(jsonObject);
-
-		facet.setFacetConfiguration(facetConfiguration);
-
-		facet.setFieldName(CPField.BASE_PRICE);
-
-		return facet;
 	}
 
 	protected String getPaginationStartParameterName(
