+++ modules/apps/commerce/commerce-product-content-search-web/src/main/java/com/liferay/commerce/product/content/search/web/internal/portlet/CPSortPortlet.java
@@ -14,33 +14,20 @@
 
 package com.liferay.commerce.product.content.search.web.internal.portlet;
 
-import com.liferay.commerce.product.constants.CPField;
 import com.liferay.commerce.product.constants.CPPortletKeys;
 import com.liferay.commerce.product.content.render.list.CPContentListRendererRegistry;
 import com.liferay.commerce.product.content.render.list.entry.CPContentListEntryRendererRegistry;
 import com.liferay.commerce.product.content.search.web.internal.display.context.CPSearchResultsDisplayContext;
 import com.liferay.commerce.product.type.CPTypeServicesTracker;
 import com.liferay.commerce.product.util.CPDefinitionHelper;
-import com.liferay.petra.string.StringBundler;
-import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.dao.search.SearchContainer;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.module.configuration.ConfigurationException;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
-import com.liferay.portal.kernel.search.Field;
-import com.liferay.portal.kernel.search.QueryConfig;
-import com.liferay.portal.kernel.search.SearchContext;
-import com.liferay.portal.kernel.search.Sort;
-import com.liferay.portal.kernel.search.SortFactoryUtil;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.portal.search.web.portlet.shared.search.PortletSharedSearchContributor;
 import com.liferay.portal.search.web.portlet.shared.search.PortletSharedSearchRequest;
 import com.liferay.portal.search.web.portlet.shared.search.PortletSharedSearchResponse;
-import com.liferay.portal.search.web.portlet.shared.search.PortletSharedSearchSettings;
 
 import java.io.IOException;
 
@@ -77,63 +64,9 @@ import org.osgi.service.component.annotations.Reference;
 		"javax.portlet.resource-bundle=content.Language",
 		"javax.portlet.security-role-ref=guest,power-user,user"
 	},
-	service = {Portlet.class, PortletSharedSearchContributor.class}
+	service = Portlet.class
 )
-public class CPSortPortlet
-	extends MVCPortlet implements PortletSharedSearchContributor {
-
-	@Override
-	public void contribute(
-		PortletSharedSearchSettings portletSharedSearchSettings) {
-
-		SearchContext searchContext =
-			portletSharedSearchSettings.getSearchContext();
-
-		QueryConfig queryConfig = portletSharedSearchSettings.getQueryConfig();
-
-		queryConfig.setHighlightEnabled(false);
-
-		RenderRequest renderRequest =
-			portletSharedSearchSettings.getRenderRequest();
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)renderRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		HttpServletRequest httpServletRequest =
-			_portal.getOriginalServletRequest(themeDisplay.getRequest());
-
-		String portletId = ParamUtil.getString(httpServletRequest, "p_p_id");
-
-		String orderByCol = ParamUtil.getString(
-			httpServletRequest,
-			StringBundler.concat(
-				StringPool.UNDERLINE, portletId, StringPool.UNDERLINE,
-				SearchContainer.DEFAULT_ORDER_BY_COL_PARAM));
-
-		if (orderByCol.equals("price-low-to-high")) {
-			searchContext.setSorts(
-				SortFactoryUtil.create(
-					CPField.BASE_PRICE, Sort.DOUBLE_TYPE, false));
-		}
-		else if (orderByCol.equals("price-high-to-low")) {
-			searchContext.setSorts(
-				SortFactoryUtil.create(
-					CPField.BASE_PRICE, Sort.DOUBLE_TYPE, true));
-		}
-		else if (orderByCol.equals("name-ascending")) {
-			searchContext.setSorts(SortFactoryUtil.create(Field.NAME, false));
-		}
-		else if (orderByCol.equals("name-descending")) {
-			searchContext.setSorts(SortFactoryUtil.create(Field.NAME, true));
-		}
-		else if (orderByCol.equals("new-items")) {
-			searchContext.setSorts(
-				SortFactoryUtil.create(Field.CREATE_DATE + "_sortable", true));
-		}
-		else {
-			searchContext.setSorts(SortFactoryUtil.getDefaultSorts());
-		}
-	}
+public class CPSortPortlet extends MVCPortlet {
 
 	@Override
 	public void render(
